Fix yield(), remove unnecessary code
​
367 / 5.000
Resultados de tradução
Resultado da tradução
A modification was made to remove yield, it will only be removed if you are using sodium, because it has a "wait for gpu" system, so it is not necessary to use yield() on the client thread in the middle of rendering. Yield() was placed in vanilla probably to prevent the client thread from going too far ahead of the GPU, this problem was solved by the sodium mod.